feat: migrate admin pages to trpc
